requirements.txt
flask==3.0.3
gunicorn==22.0.0
openai==1.40.0
google-auth==2.33.0
requests==2.32.3
